remove verify script
